Rename opencl.h to worker.h and same for .c files
