+++ src/worker/worker.h
