+++ test/benchmark.c
@@ -22,7 +22,7 @@
         ((t1_ptr)->tv_sec-(t0_ptr)->tv_sec)*1000+((t1_ptr)->tv_nsec-(t0_ptr)->tv_nsec)/1000000
 #endif
 
-#include "../src/worker/opencl.h"
+#include "../src/worker/worker.h"
 #include "../src/mine_DUCO_S1.h"
 #include "../src/mine_xxhash.h"
